refactor(coursition): rename component files to kebab-case
