+++ apps/coursition/app/components/status-display.tsx
